Move DigdagEmbed and LocalSite to digdag-server. colect bootstrap codes to digdag-server.
