+++ digdag-cli/src/main/java/io/digdag/cli/Run.java
@@ -13,9 +13,9 @@ import com.google.inject.multibindings.Multibinder;
 import io.digdag.client.config.Config;
 import io.digdag.client.config.ConfigException;
 import io.digdag.client.config.ConfigFactory;
-import io.digdag.executor.DigdagEmbed;
-import io.digdag.executor.LocalSite;
-import io.digdag.executor.LocalSite.StoreWorkflowResult;
+import io.digdag.server.DigdagEmbed;
+import io.digdag.server.LocalSite;
+import io.digdag.server.LocalSite.StoreWorkflowResult;
 import io.digdag.core.agent.AgentConfig;
 import io.digdag.core.agent.AgentId;
 import io.digdag.core.agent.ConfigEvalEngine;
