+++ digdag-cli/src/main/java/io/digdag/cli/client/Archive.java
@@ -10,7 +10,7 @@ import io.digdag.cli.StdOut;
 import io.digdag.cli.SystemExitException;
 import io.digdag.cli.YamlMapper;
 import io.digdag.client.config.ConfigElement;
-import io.digdag.executor.DigdagEmbed;
+import io.digdag.server.DigdagEmbed;
 
 import java.io.IOException;
 import java.io.PrintStream;
