+++ digdag-cli/src/main/java/io/digdag/cli/client/Start.java
@@ -18,7 +18,7 @@ import io.digdag.client.api.SessionTimeTruncate;
 import io.digdag.client.config.Config;
 import io.digdag.client.config.ConfigElement;
 import io.digdag.client.config.ConfigFactory;
-import io.digdag.executor.DigdagEmbed;
+import io.digdag.server.DigdagEmbed;
 import io.digdag.core.config.ConfigLoaderManager;
 
 import javax.ws.rs.ClientErrorException;
