+++ digdag-core/src/test/java/io/digdag/core/workflow/WorkflowTestingUtils.java
@@ -7,8 +7,8 @@ import com.google.inject.Scopes;
 import com.google.inject.multibindings.Multibinder;
 import io.digdag.client.config.Config;
 import io.digdag.client.config.ConfigUtils;
-import io.digdag.executor.DigdagEmbed;
-import io.digdag.executor.LocalSite;
+import io.digdag.server.DigdagEmbed;
+import io.digdag.server.LocalSite;
 import io.digdag.core.config.YamlConfigLoader;
 import io.digdag.core.crypto.SecretCrypto;
 import io.digdag.core.crypto.SecretCryptoProvider;
