+++ digdag-standards/src/test/java/io/digdag/standards/operator/ForEachOperatorFactoryTest.java
@@ -9,7 +9,7 @@ import org.junit.rules.TemporaryFolder;
 import java.nio.file.Path;
 
 import io.digdag.client.config.Config;
-import io.digdag.executor.DigdagEmbed;
+import io.digdag.server.DigdagEmbed;
 import io.digdag.spi.TaskResult;
 import io.digdag.standards.operator.ForEachOperatorFactory.ForEachOperator;
