+++ digdag-standards/src/test/java/io/digdag/standards/operator/ForRangeOperatorFactoryTest.java
@@ -1,7 +1,7 @@
 package io.digdag.standards.operator;
 
 import io.digdag.client.config.Config;
-import io.digdag.executor.DigdagEmbed;
+import io.digdag.server.DigdagEmbed;
 import io.digdag.spi.TaskResult;
 import io.digdag.standards.operator.ForRangeOperatorFactory.ForRangeOperator;
 import java.nio.file.Path;
